Fix paeth predictor to build without SIMD
